Search only on button click
This hides a multitude of sins.  It would be nice to have everything working so tightly that we could do as-you-type searching, but there are lots of other fish to fry, and this neatly sidesteps a host of potentially really tricky problems (for now, at least) at fairly minimal cost.