Rewrite feature cropping
This is a rewrite resulting from multiple different experiments using workers, rust or libvips in wasm, offscreen canvases, and many other approaches.  The code here is lacking polish but is the most performant I've been able to come up with so far (really it's the only version that can handle thousands of features across multiple very large images), so this commit is a baseline for further finessing and UI niceties.